UBUNTU: SAUCE: (drop after 3.2) ALSA: hda - restrict bass configuration on Dell Inspiron 17
The ALC269VB configuration on Dell Inspiron 17 results in mixer volume
controls only for the main and "bass"/surround speakers, lacking a
separate volume control for headphones, due to the lack of DACs
available. Realtek auto config avoids to create an duplicate volume
slider for the same nid (DAC), so as headphone and first set of speakers
share the same DAC, and the volume control is done on DAC amp, only one
volume slider with the name "Speaker Playback Volume" is created.

The problem is that on Ubuntu pulse configuration also uses control
names to do automute. So when you insert a headphone, it goes there and
mutes and lowers volumes of all Speaker controls. As Speaker volume
slider also controls the volume on headphones, this ends up also
"muting" the headphone output, and you have bug 994685 reported.

This change workarounds the issue, forcing the current auto config code
in 3.2 kernel to allocate an exclusive DAC for headphones, which allows
the creation of separate Headphone and Speaker controls. We do that
forcing the headphone pins and speaker pins to "see" only the mixers
routing to DACs we want, and after the setup is done we undo the hidden
connections.

Also we make sure to add a mixer item back to control the second speaker
pin, otherwise it is gone with the workaround in place.

And as it happens sometimes, where manufactures end up using their same
subvendor and subdevice ids for different setups/hardware, we play safe
and do some sanity checks first before applying the fixup.

BugLink: http://bugs.launchpad.net/bugs/994685
Signed-off-by: Herton Ronaldo Krzesinski <herton.krzesinski@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>