UBUNTU: add support for generating binary device trees and install them in /lib/firmware
Add support for generating a device tree binary for this flavour and
installing it in /lib/firmware.  This is useful for QEMU for some flavours
and likely will be useful later for systems which need to include a device
tree when flashing the kernel (this is the overall vision for device tree
with generic ARM kernels).

Adds flavour specific configuration for the name of the dtb_file.
Currently this only supports one dtb file per flavour but it is possible
there could be more than one when flavours can support more than one SOC,
so futureproof by including the hardware name in the dtb path.

Based on a patch by Dann Frazier.

BugLink: http://bugs.launchpad.net/bugs/1030600
Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Herton Krzesinski <herton.krzesinski@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>